Reload config based on mtime (#1043)
