+++ packages/zudoku/src/config/file-exists.ts
@@ -1,0 +1,6 @@
+import { stat } from "node:fs/promises";
+
+export const fileExists = (path: string) =>
+  stat(path)
+    .then(() => true)
+    .catch(() => false);
