+++ packages/zudoku/src/lib/plugins/markdown/index.tsx
@@ -1,8 +1,8 @@
 import type { Toc } from "@stefanprobst/rehype-extract-toc";
 import type { MDXProps } from "mdx/types.js";
 import { type JSX } from "react";
-import { RouteObject } from "react-router";
-import { ZudokuDocsConfig } from "../../../config/validators/common.js";
+import type { RouteObject } from "react-router";
+import type { ZudokuDocsConfig } from "../../../config/validators/validate.js";
 import type { ZudokuPlugin } from "../../core/plugins.js";
 import { DocResolver } from "./resolver.js";
