+++ packages/zudoku/src/zuplo/with-zuplo.ts
@@ -1,9 +1,9 @@
-import type { CommonConfig } from "../config/validators/common.js";
+import type { ZudokuConfig } from "../config/validators/validate.js";
 
 /**
  *  @deprecated `withZuplo` is no longer needed and will automatically be applied. It will be removed in a future version.
  */
-const withZuplo = <TConfig extends CommonConfig>(config: TConfig): TConfig => {
+const withZuplo = (config: ZudokuConfig): ZudokuConfig => {
   // eslint-disable-next-line no-console
   console.warn(
     "`withZuplo` is no longer needed and will automatically be applied. It will be removed in a future version.",
